Plot range and label size updates to histPlot
This updates the histPlot action to allow for finer grained control
over the plot ranges.  The "rangeType" can now be selected as one of
"percentile", "sigmaMad", or "fixed".  The lower and upper bounds of
the range will then be set accordingly by the values in lowerRange and
upperRange as follows:

"percentile": (lowerRange percentile of data,
               upperRange percentile of data)

"sigmaMad": (min(medians) - lowerRange*max(sigmaMads),
             max(medians) + upperRange*max(sigmaMads))

"fixed": (lowerRange, upperRange)

It also updates the way the right-hand statistics legends are added
such that each entry gets a title (set to the x-axis label) and the
legends line up (reasonably well for up to ~6 panels with <~ 4
datasets per panel) with their respective rows.

In order to avoid confusion between col for color vs. column, all
references to col refer to the latter, and color is spelled out for
the former.