Update the SkyObject[Source]HistPlot classes
This updates the skyObject and skySource analysisPlot classes to
adapt to the new range setting configs in the HistPlot action.
This also adds vertical solid black lines at a "reference" value
(0.0 here) in both the flux and S/N panels as well as a shaded
PDF(0, 1) distribution on the now density normalized S/N plots
for reference.

This also adds the GaaP 1p0 flux to the skyObject histograms (GaaP
fluxes are not measured for the visit-level sky sources).