Structural changes to layout and function
This is a catch-all commit for many structural changes to the
package. This covers unifying analysisPlots and analysisMetrics
as well as laying the ground work for dispatching Metrics to
a time series database. There are a few other quality of life
changes that were made during the refactoring process such as
the introduction of a finalize method.