Add ability to dispatch metrics to Sasquatch
This commit adds the ability to dispatch MetricMeasurementBundles
to the Sasquatch service. This can be done either with the
SasquatchDisaptcher class, or more commonly using a butler
datastore that will automatically dispatch them when put to the
datastore. There are a few other cleanups on this commit that were
made either to support this work, or that were found in the
process of doing this ticket.