Fix mypy configuration.
The previous configuration led to mypy doing nothing, because it was
configured to ignore all errors outside of lsst.daf.butler (hmm, I
wonder where the config file came from originally?).

As one would expect, fixing this leads to a lot of mypy errors, and
for now I've just told mypy to ignore the files that have them.