Add accommodations for PSF robustsness metric
The ccdVisitTable now include a column named psfTraceRadiusDelta that
provides a measure of the robustness of the PSF model.  This metric is
thresholded against inclusion in the coadds (in makeWarp), so is useful
to track visually (as with the other thresholded metrics).  Add the
accommotations required for this quantity in the multiVisitCoverage
plot action.