Change contours to dist vs. color in stellarLocus plot
It was deemed that a distance from fit vs. color scatter plot (showing
a running median) was much more informative (i.e. would highlight any
potential color dependency/bias in the fit) than the magnitude vs.
distance to fit contour plot, so this is added here are set as the
default plot for the lower right panel. The original contours plot
can be selected instead by setting the doPlotDistVsColor config to
False.