Add magnitude range selector
The use case here is for the stellar locus data selection. While we
still rely on a S/N threshold, we also want the analysis to always
include the same magnitude range, regardless of the depth of any
given image, to minimize differences in the stellar populations
included in the stellar locus analysis. (This would break down for
abnormally shallow or low S/N images, so this should be kept in mind
when setting thresholds and assessing results.)