Fix breakage in BasePrep
The ticket adding the Tensor interface make a change to BasePrep
to make it more logically sound to the types that are possible in
keyed data. This unfortunately overlooked usage that existed that
relied on ill defined behavior. This is a fixup to that, restoring
existing behavior is a slightly more gaurded way.