Make BasePrep.addInputSchema idempotent
Add input schema was adding the same vectors multiple times due to
how serialization was happening with pickle. Make this idempotent
across the number of times this method is called. This is also a
good change to make to guard against mistakes by the user specifying
a vector multiple times.