Handle originally-ArrowAstropy objectTables in source match.
This changes ObjectEpochTableTask to read objectTable in directly as
ArrowAstropy, because it's easy.  It does not change
SourceObjectMatch to read in objectTable as ArrowAstropy because it's
hard - there are pandas-specific operations, specially pd.merge, that
I don't want to try to work out how to do natively in Astropy.  Instead
we make sure the DataFrame that's read doesn't rely on having the
index set to the objectId, since that's what goes away when the dataset
is originally written as ArrowAstropy.