Allow raising NoWorkFound in metadata metrics tasks
If NoWorkFound was raised in the original PipelineTask, it will still write empty metadata.
The metadata parsing task will then fail, which will incorrectly be reported as a pipeline failure.
Instead allow raising NoWorkFound in cases where none of the configured metrics are found.

Note that this does not work correctly with metrics in subtasks, but as long as any metric is found it will not raise NoWorkFound