WIP: add option to raise NoWorkFound if any metadata is missing.
This adds a config option to all MetadataAnalysisTasks even though
only TaskMetadataAnalysisTask uses it; this is the same (bad) pattern
used for the existing raiseNoWorkFoundOnEmptyMetadata option.  But
right now it's unclear whether the right fix is to handle it in both
derived classes or move the config options to a derived config.