Use finalVisitSummary in concrete CatalogMatchVisit tasks.
The visit summary is only used to get the observation date (which
doesn't change when we make the final visit summary), so this wasn't
doing anything wrong before.  But in the future we might want to
delete the initial visit summary as soon as the final one is available,
or someone might add code to these tasks to use more than just the
date, so using the final visit summary is more consistent with the
rest of the defaults (e.g. sourceTable_visit as the input catalog).

Configurations that use preSourceTable_visit as the input catalog will
be adjusted downstream in drp_pipe to switch back to visitSummary
instead of finalVisitSummary.