Include outputName prefix in run-repeatedly task configs.
If we have a tool configuration in a separate config file, it's
because we expect to run those configurations multiple times (with
different connections and output names) in the same pipeline.  And if
we run something multiple times in the same pipeline, we need
something (like the outputName) in the dispatched metric name to avoid
conflicts.