Catch errors when running atools and finally raise
Here we catch errors when attempting to run an analysis tools action and
continue if an error is triggered.
Finally, an AnnotatedPartialOutputsError is raised containing
information on each reported error.
This is almost certainly not the correct thing to do here, as the
analysis tool will still fail with the raised error.
Perhaps what we want instead is something more akin to NoWorkFound, but
which remains valid when partial work has been found (e.g.,
PartialWorkFound might be a suitable name).
This then allows subsequent downstream processing to continue when only
partial inputs were provided and partial outputs generated.