Reconcile low stats threshold in scatterPlotWithTwoHists
The minimum number of xs to calculate had previously been 2.
A subsequent edit changed this to 10 in one place, but
didn't update the value from 2 later in the code. To avoid
this happening in the future, I've added a min_n_xs_for_stats
variable, which I have set to 10.